refactor(tangle-dapp): Move blueprints listing API to shared ui
