+++ apps/tangle-dapp/app/blueprints/BlueprintListing.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import BlueprintGallery from '@webb-tools/tangle-shared-ui/components/blueprints/BlueprintGallery';
+import useBlueprintListing from '@webb-tools/tangle-shared-ui/data/blueprints/useBlueprintListing';
 import Image from 'next/image';
 import { FC } from 'react';
 
-import useBlueprintListing from '../../data/blueprints/useBlueprintListing';
 import { PagePath } from '../../types';
 
 const BlueprintListing: FC = () => {
