+++ apps/tangle-dapp/app/restake/page.tsx
@@ -7,6 +7,7 @@ import {
   UserLineIcon,
 } from '@webb-tools/icons';
 import { OnboardingPageKey } from '@webb-tools/tangle-shared-ui/constants';
+import useRestakeOperatorMap from '@webb-tools/tangle-shared-ui/data/restake/useRestakeOperatorMap';
 import { TANGLE_DOCS_RESTAKING_URL } from '@webb-tools/webb-ui-components';
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
 import { Typography } from '@webb-tools/webb-ui-components/typography/Typography';
@@ -17,7 +18,6 @@ import OnboardingItem from '../../components/OnboardingModal/OnboardingItem';
 import OnboardingModal from '../../components/OnboardingModal/OnboardingModal';
 import StatItem from '../../components/StatItem';
 import useRestakeDelegatorInfo from '../../data/restake/useRestakeDelegatorInfo';
-import useRestakeOperatorMap from '../../data/restake/useRestakeOperatorMap';
 import useRestakeTVL from '../../data/restake/useRestakeTVL';
 import getTVLToDisplay from '../../utils/getTVLToDisplay';
 import TableTabs from './TableTabs';
