+++ apps/tangle-dapp/utils/getAmountValidation.ts
@@ -1,7 +1,6 @@
 import { ZERO_BIG_INT } from '@webb-tools/dapp-config/constants';
+import safeParseUnits from '@webb-tools/tangle-shared-ui/utils/safeParseUnits';
 import Decimal from 'decimal.js';
-
-import safeParseUnits from './safeParseUnits';
 
 /**
  * Get the shared amount validation for react-hook-form
