+++ libs/tangle-shared-ui/src/data/blueprints/utils/toPrimitiveBlueprint.ts
